Update README.md
This is a re-upload of the Chrome web store extension "Share to Facebook" released by Facebook.

It has been retracted for some reason, and i like the Quote function it provides, which i haven't found on any other extensions. 

Simply load this zip file in developer mode in your Chrome browser.